Added dataset with Post Code information from CBS
I am adding 3 datasets to the fit repo that contain data such as the WOZ value of houses per postcode. The three datasets contain the same variables but the postcodes are a bit different. pc6 is the dataset for the full postcode with 6 letters. Such a postcode is unique to a street. Every street has its own post code. CBS has a privacy protection clause which says that is there is data of 4 or less people or houses in a postcode pc4, pc5 or pc6, they will not show the data, so you see that in the pc6 dataset, a lot of the data is missing. Missing data is represented by the -99997 number.