Added coordinates for each address
I have added a code block that uses the Google Geocoding API to extract the coordinates of the addresses. This can later be used to collect the information about the dtstance to amenities variable. This could also be used to check how many amenities such as cafes are within a 100m or 1000m radius