Added code for number of amenities in 100m and 1000m radius
I have been able to use the Places API to add the amount of amenities within 100m and 1000m, but for 10 listins it took 3 minutes and a couple of seconds. For 900 listings, the processing time would be insane. I am considering using the cloud to process all of this. For this, I will need to finish all the other new variable code as well so I can run it at the same time.