Imputed missing values with kNN Imputer
I checked everything and the imputation looks good except for thetoilet one. bathroom = 1 means that there is a shower and toilet in 1 room. Toilet = 1 is a separate toilet from the shower. It is not completely correct and I am no sure how I would be able to correct it. I might have to drop the column bathroom and toilet, but I'll do some research into how it is viewed in Funda, cause it could be that toilet will always be included in bathroom variable. In that case, we can leave bathroom and drop toilet