Chose transformations and did PCA
I have chosen the transformations based on the linearity using pearson correlation. Rent_log gave the best overall performance. I performed PCA on all the spatial data and it gave me 3 components that explain 85% of the variance, which is great. I saw that there were some outliers in the train set, so I will check how I will handle this. There is also heteroscedasticity in the residuals so that is also an issue. If I can solve those, I can start with the actual fitting. I already fitted the RFR and LR and predicted the values. Surprisingly, the LR and RFR perform very similarly with RFR (not optimized yet) a bit better than LR.